# aula 7/doc/openapi.yaml
openapi: 3.0.3

info:
  title: Projeto controle de filmes - SENAI
  description: API destinada ao controle de filmes para o projeto "Bem Legal" do SENAI
  version: 1.0.0

servers:
  - url: http://localhost:3030

paths:

  /v1/senai/locadora/filme:

    post:
      tags:
        - Filme

      summary: Adiciona um novo filme

      description: Endpoint para enviar os dados do filme a serem cadastrados

      requestBody:
        description: JSON contendo os dados do novo filme
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filme'

      responses:
        '201':
          description: Filme cadastrado com sucesso

        '400':
          description: Invalid input

        '415':
          description: Invalid content-type

    get:
      tags:
        - Filme

      summary: Lista todos os filmes

      description: Listagem de todos os filmes registrados no sistema

      responses:
        '200':
          description: Successful operation

          content:
            application/json:
              schema:
                type: array

                items:
                  $ref: '#/components/schemas/filme'

        '404':
          description: Not found

  /v1/senai/locadora/filme/{id}:

    get:
      tags:
        - Filme

      summary: Busca um filme pelo ID

      description: Retorna um filme pelo ID

      parameters:
        - name: id
          in: path
          required: true

          description: ID do filme

          schema:
            type: integer

      responses:
        '200':
          description: Successful operation

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/filme'

        '400':
          description: Invalid ID supplied

        '404':
          description: Not found

        '500':
          description: Error internal server

    put:
      tags:
        - Filme

      summary: Atualiza um filme pelo ID

      description: Atualiza um filme existente

      parameters:
        - name: id
          in: path
          required: true

          description: ID do filme

          schema:
            type: integer

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filme'

      responses:
        '200':
          description: Filme atualizado com sucesso

        '400':
          description: Invalid ID supplied

        '404':
          description: Not found

        '415':
          description: Invalid content-type

        '500':
          description: Error internal server

    delete:
      tags:
        - Filme

      summary: Exclui um filme pelo ID

      description: Remove um filme do sistema

      parameters:
        - name: id
          in: path
          required: true

          description: ID do filme

          schema:
            type: integer

      responses:
        '200':
          description: Filme deletado com sucesso

        '400':
          description: Invalid ID supplied

        '404':
          description: Not found

        '500':
          description: Error internal server

components:

  schemas:

    filme:
      type: object

      required:
        - nome
        - sinopse
        - data_lancamento

      properties:

        id:
          type: integer
          example: 1

        nome:
          type: string
          example: Vingadores Ultimato

        sinopse:
          type: string
          example: Os heróis restantes enfrentam Thanos.

        capa:
          type: string
          example: https://imagem.com/capa.jpg

        valor:
          type: number
          example: 29.90

        data_lancamento:
          type: string
          example: 2019-04-25

        avaliacao:
          type: number
          example: 9.5